Add r/w utility function for configuration datastore
NOTE: This patch moved the path manager model from operational to
ocnfiguration. The updater will still pdate the operational datastore,
but it will not be used by the unicorn impl. This is a temp solution
to support configured external path provider. We still need to
implement the sync from configuration to operational datastore later.

Signed-off-by: jensenzhang <jingxuan.n.zhang@gmail.com>